move bundling logic to sandbox.ts
